+++ glide-core/redis-rs/redis/src/cluster_async/mod.rs
@@ -845,6 +845,7 @@ impl<C> Future for Request<C> {
                 let request = this.request.as_mut().unwrap();
                 // TODO - would be nice if we didn't need to repeat this code twice, with & without retries.
                 if request.retry >= this.retry_params.number_of_retries {
+                    let retry_method = err.retry_method();
                     let next = if err.kind() == ErrorKind::AllConnectionsUnavailable {
                         Next::ReconnectToInitialNodes { request: None }.into()
                     } else if matches!(err.retry_method(), crate::types::RetryMethod::MovedRedirect)
@@ -855,7 +856,9 @@ impl<C> Future for Request<C> {
                             sleep_duration: None,
                         }
                         .into()
-                    } else if matches!(err.retry_method(), crate::types::RetryMethod::Reconnect) {
+                    } else if matches!(retry_method, crate::types::RetryMethod::Reconnect)
+                        || matches!(retry_method, crate::types::RetryMethod::ReconnectAndRetry)
+                    {
                         if let OperationTarget::Node { address } = target {
                             Next::Reconnect {
                                 request: None,
@@ -934,13 +937,18 @@ impl<C> Future for Request<C> {
                         });
                         self.poll(cx)
                     }
-                    crate::types::RetryMethod::Reconnect => {
+                    crate::types::RetryMethod::Reconnect
+                    | crate::types::RetryMethod::ReconnectAndRetry => {
                         let mut request = this.request.take().unwrap();
                         // TODO should we reset the redirect here?
                         request.info.reset_routing();
                         warn!("disconnected from {:?}", address);
+                        let should_retry = matches!(
+                            err.retry_method(),
+                            crate::types::RetryMethod::ReconnectAndRetry
+                        );
                         Next::Reconnect {
-                            request: Some(request),
+                            request: should_retry.then_some(request),
                             target: address,
                         }
                         .into()
@@ -1177,8 +1185,11 @@ where
         Ok(connections.0)
     }
 
-    fn reconnect_to_initial_nodes(&mut self) -> impl Future<Output = ()> {
-        let inner = self.inner.clone();
+    // Reconnet to the initial nodes provided by the user in the creation of the client,
+    // and try to refresh the slots based on the initial connections.
+    // Being used when all cluster connections are unavailable.
+    fn reconnect_to_initial_nodes(inner: Arc<InnerCore<C>>) -> impl Future<Output = ()> {
+        let inner = inner.clone();
         async move {
             let connection_map = match Self::create_initial_connections(
                 &inner.initial_nodes,
@@ -1207,7 +1218,7 @@ where
         }
     }
 
-    // Validate all existing user connections and try to reconnect if nessesary.
+    // Validate all existing user connections and try to reconnect if necessary.
     // In addition, as a safety measure, drop nodes that do not have any assigned slots.
     // This function serves as a cheap alternative to slot_refresh() and thus can be used much more frequently.
     // The function does not discover the topology from the cluster and assumes the cached topology is valid.
@@ -1282,7 +1293,7 @@ where
         let connections_container = inner.conn_lock.read().await;
         let cluster_params = &inner.cluster_params;
         let subscriptions_by_address = &inner.subscriptions_by_address;
-        let glide_connection_optons = &inner.glide_connection_options;
+        let glide_connection_options = &inner.glide_connection_options;
 
         stream::iter(addresses.into_iter())
             .fold(
@@ -1304,7 +1315,7 @@ where
                         node_option,
                         &cluster_params,
                         conn_type,
-                        glide_connection_optons.clone(),
+                        glide_connection_options.clone(),
                     )
                     .await;
                     match node {
@@ -1517,7 +1528,7 @@ where
                 Self::check_topology_and_refresh_if_diff(inner.clone(), &RefreshPolicy::Throttable)
                     .await;
             if !topology_changed {
-                // This serves as a safety measure for validating pubsub subsctiptions state in case it has drifted
+                // This serves as a safety measure for validating pubsub subscriptions state in case it has drifted
                 // while topology stayed the same.
                 // For example, a failed attempt to refresh a connection which is triggered from refresh_pubsub_subscriptions(),
                 // might leave a node unconnected indefinitely in case topology is stable and no request are attempted to this node.
@@ -1680,7 +1691,9 @@ where
         Self::refresh_slots_inner(inner, curr_retry)
             .await
             .map_err(|err| {
-                if curr_retry > DEFAULT_NUMBER_OF_REFRESH_SLOTS_RETRIES {
+                if curr_retry > DEFAULT_NUMBER_OF_REFRESH_SLOTS_RETRIES
+                    || err.kind() == ErrorKind::AllConnectionsUnavailable
+                {
                     BackoffError::Permanent(err)
                 } else {
                     BackoffError::from(err)
@@ -2073,14 +2086,22 @@ where
             }
             ConnectionCheck::RandomConnection => {
                 let read_guard = core.conn_lock.read().await;
-                let (random_address, random_conn_future) = read_guard
+                read_guard
                     .random_connections(1, ConnectionType::User)
-                    .next()
-                    .ok_or(RedisError::from((
-                        ErrorKind::AllConnectionsUnavailable,
-                        "No random connection found",
-                    )))?;
-                return Ok((random_address, random_conn_future.await));
+                    .and_then(|mut random_connections| {
+                        random_connections.next().map(
+                            |(random_address, random_conn_future)| async move {
+                                (random_address, random_conn_future.await)
+                            },
+                        )
+                    })
+                    .ok_or_else(|| {
+                        RedisError::from((
+                            ErrorKind::AllConnectionsUnavailable,
+                            "No random connection found",
+                        ))
+                    })?
+                    .await
             }
         };
 
@@ -2104,10 +2125,19 @@ where
                 }
                 Err(err) => {
                     trace!("Recover slots failed!");
-                    *future = Box::pin(Self::refresh_slots_and_subscriptions_with_retries(
-                        self.inner.clone(),
-                        &RefreshPolicy::Throttable,
-                    ));
+                    let next_state = if err.kind() == ErrorKind::AllConnectionsUnavailable {
+                        ConnectionState::Recover(RecoverFuture::Reconnect(Box::pin(
+                            ClusterConnInner::reconnect_to_initial_nodes(self.inner.clone()),
+                        )))
+                    } else {
+                        ConnectionState::Recover(RecoverFuture::RecoverSlots(Box::pin(
+                            Self::refresh_slots_and_subscriptions_with_retries(
+                                self.inner.clone(),
+                                &RefreshPolicy::Throttable,
+                            ),
+                        )))
+                    };
+                    self.state = next_state;
                     Poll::Ready(Err(err))
                 }
             },
@@ -2226,9 +2256,7 @@ where
                         }));
                     }
                 }
-                Next::Reconnect {
-                    request, target, ..
-                } => {
+                Next::Reconnect { request, target } => {
                     poll_flush_action =
                         poll_flush_action.change_state(PollFlushAction::Reconnect(vec![target]));
                     if let Some(request) = request {
@@ -2371,7 +2399,7 @@ where
                 }
                 PollFlushAction::ReconnectFromInitialConnections => {
                     self.state = ConnectionState::Recover(RecoverFuture::Reconnect(Box::pin(
-                        self.reconnect_to_initial_nodes(),
+                        ClusterConnInner::reconnect_to_initial_nodes(self.inner.clone()),
                     )));
                 }
             }
@@ -2413,8 +2441,19 @@ async fn calculate_topology_from_random_nodes<'a, C>(
 where
     C: ConnectionLike + Connect + Clone + Send + Sync + 'static,
 {
-    let requested_nodes =
-        read_guard.random_connections(num_of_nodes_to_query, ConnectionType::PreferManagement);
+    let requested_nodes = if let Some(random_conns) =
+        read_guard.random_connections(num_of_nodes_to_query, ConnectionType::PreferManagement)
+    {
+        random_conns
+    } else {
+        return (
+            Err(RedisError::from((
+                ErrorKind::AllConnectionsUnavailable,
+                "No available connections to refresh slots from",
+            ))),
+            vec![],
+        );
+    };
     let topology_join_results =
         futures::future::join_all(requested_nodes.map(|(addr, conn)| async move {
             let mut conn: C = conn.await;