+++ java/client/src/main/java/glide/api/commands/servermodules/Json.java
@@ -20,12 +20,16 @@ import lombok.NonNull;
 public class Json {
 
     private static final String JSON_PREFIX = "JSON.";
-    public static final String JSON_SET = JSON_PREFIX + "SET";
-    public static final String JSON_GET = JSON_PREFIX + "GET";
+    private static final String JSON_SET = JSON_PREFIX + "SET";
+    private static final String JSON_GET = JSON_PREFIX + "GET";
+    private static final String JSON_ARRAPPEND = JSON_PREFIX + "ARRAPPEND";
     private static final String JSON_ARRINSERT = JSON_PREFIX + "ARRINSERT";
     private static final String JSON_ARRLEN = JSON_PREFIX + "ARRLEN";
     private static final String JSON_NUMINCRBY = JSON_PREFIX + "NUMINCRBY";
     private static final String JSON_NUMMULTBY = JSON_PREFIX + "NUMMULTBY";
+    private static final String JSON_DEL = JSON_PREFIX + "DEL";
+    private static final String JSON_FORGET = JSON_PREFIX + "FORGET";
+    private static final String JSON_TOGGLE = JSON_PREFIX + "TOGGLE";
 
     private Json() {}
 
@@ -42,7 +46,7 @@ public class Json {
      * @return A simple <code>"OK"</code> response if the value is successfully set.
      * @example
      *     <pre>{@code
-     * String value = Json.set(client, "doc", ".", "{'a': 1.0, 'b': 2}").get();
+     * String value = Json.set(client, "doc", ".", "{\"a\": 1.0, \"b\": 2}").get();
      * assert value.equals("OK");
      * }</pre>
      */
@@ -67,7 +71,7 @@ public class Json {
      * @return A simple <code>"OK"</code> response if the value is successfully set.
      * @example
      *     <pre>{@code
-     * String value = Json.set(client, gs("doc"), gs("."), gs("{'a': 1.0, 'b': 2}")).get();
+     * String value = Json.set(client, gs("doc"), gs("."), gs("{\"a\": 1.0, \"b\": 2}")).get();
      * assert value.equals("OK");
      * }</pre>
      */
@@ -94,7 +98,7 @@ public class Json {
      *     set because of <code>setCondition</code>, returns <code>null</code>.
      * @example
      *     <pre>{@code
-     * String value = Json.set(client, "doc", ".", "{'a': 1.0, 'b': 2}", ConditionalChange.ONLY_IF_DOES_NOT_EXIST).get();
+     * String value = Json.set(client, "doc", ".", "{\"a\": 1.0, \"b\": 2}", ConditionalChange.ONLY_IF_DOES_NOT_EXIST).get();
      * assert value.equals("OK");
      * }</pre>
      */
@@ -123,7 +127,7 @@ public class Json {
      *     set because of <code>setCondition</code>, returns <code>null</code>.
      * @example
      *     <pre>{@code
-     * String value = Json.set(client, gs("doc"), gs("."), gs("{'a': 1.0, 'b': 2}"), ConditionalChange.ONLY_IF_DOES_NOT_EXIST).get();
+     * String value = Json.set(client, gs("doc"), gs("."), gs("{\"a\": 1.0, \"b\": 2}"), ConditionalChange.ONLY_IF_DOES_NOT_EXIST).get();
      * assert value.equals("OK");
      * }</pre>
      */
@@ -148,7 +152,7 @@ public class Json {
      * @example
      *     <pre>{@code
      * String value = Json.get(client, "doc").get();
-     * assert value.equals("{'a': 1.0, 'b': 2}");
+     * assert value.equals("{\"a\": 1.0, \"b\": 2}");
      * }</pre>
      */
     public static CompletableFuture<String> get(@NonNull BaseClient client, @NonNull String key) {
@@ -165,7 +169,7 @@ public class Json {
      * @example
      *     <pre>{@code
      * GlideString value = Json.get(client, gs("doc")).get();
-     * assert value.equals(gs("{'a': 1.0, 'b': 2}"));
+     * assert value.equals(gs("{\"a\": 1.0, \"b\": 2}"));
      * }</pre>
      */
     public static CompletableFuture<GlideString> get(
@@ -200,7 +204,7 @@ public class Json {
      * @example
      *     <pre>{@code
      * String value = Json.get(client, "doc", new String[] {"$"}).get();
-     * assert value.equals("{'a': 1.0, 'b': 2}");
+     * assert value.equals("{\"a\": 1.0, \"b\": 2}");
      * String value = Json.get(client, "doc", new String[] {"$.a", "$.b"}).get();
      * assert value.equals("{\"$.a\": [1.0], \"$.b\": [2]}");
      * }</pre>
@@ -237,7 +241,7 @@ public class Json {
      * @example
      *     <pre>{@code
      * GlideString value = Json.get(client, gs("doc"), new GlideString[] {gs("$")}).get();
-     * assert value.equals(gs("{'a': 1.0, 'b': 2}"));
+     * assert value.equals(gs("{\"a\": 1.0, \"b\": 2}"));
      * GlideString value = Json.get(client, gs("doc"), new GlideString[] {gs("$.a"), gs("$.b")}).get();
      * assert value.equals(gs("{\"$.a\": [1.0], \"$.b\": [2]}"));
      * }</pre>
@@ -390,6 +394,86 @@ public class Json {
         return executeCommand(
                 client,
                 new ArgsBuilder().add(gs(JSON_GET)).add(key).add(options.toArgs()).add(paths).toArray());
+    }
+
+    /**
+     * Appends one or more <code>values</code> to the JSON array at the specified <code>path</code>
+     * within the JSON document stored at <code>key</code>.
+     *
+     * @param client The client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @param path Represents the <code>path</code> within the JSON document where the <code>values
+     *     </code> will be appended.
+     * @param values The <code>values</code> to append to the JSON array at the specified <code>path
+     *     </code>.
+     * @return
+     *     <ul>
+     *       <li>For JSONPath (<code>path</code> starts with <code>$</code>):<br>
+     *           Returns a list of integers for every possible path, indicating the new length of the
+     *           new array after appending <code>values</code>, or <code>null</code> for JSON values
+     *           matching the path that are not an array. If <code>path</code> does not exist, an
+     *           empty array will be returned.
+     *       <li>For legacy path (<code>path</code> doesn't start with <code>$</code>):<br>
+     *           Returns the length of the new array after appending <code>values</code> to the array
+     *           at <code>path</code>. If multiple paths are matched, returns the last updated array.
+     *           If the JSON value at <code>path</code> is not a array or if <code>path</code> doesn't
+     *           exist, an error is raised. If <code>key</code> doesn't exist, an error is raised.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", "$", "{\"a\": 1, \"b\": [\"one\", \"two\"]}").get();
+     * var res = Json.arrappend(client, "doc", "$.b", new String[] {"\"three\""}).get();
+     * assert Arrays.equals((Object[]) res, new int[] {3}); // New length of the array after appending
+     * res = Json.arrappend(client, "doc", ".b", new String[] {"\"four\""}).get();
+     * assert res.equals(4); // New length of the array after appending
+     * }</pre>
+     */
+    public static CompletableFuture<Object> arrappend(
+            @NonNull BaseClient client,
+            @NonNull String key,
+            @NonNull String path,
+            @NonNull String[] values) {
+        return executeCommand(
+                client, concatenateArrays(new String[] {JSON_ARRAPPEND, key, path}, values));
+    }
+
+    /**
+     * Appends one or more <code>values</code> to the JSON array at the specified <code>path</code>
+     * within the JSON document stored at <code>key</code>.
+     *
+     * @param client The client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @param path Represents the <code>path</code> within the JSON document where the <code>values
+     *     </code> will be appended.
+     * @param values The <code>values</code> to append to the JSON array at the specified <code>path
+     *     </code>.
+     * @return
+     *     <ul>
+     *       <li>For JSONPath (<code>path</code> starts with <code>$</code>):<br>
+     *           Returns a list of integers for every possible path, indicating the new length of the
+     *           new array after appending <code>values</code>, or <code>null</code> for JSON values
+     *           matching the path that are not an array. If <code>path</code> does not exist, an
+     *           empty array will be returned.
+     *       <li>For legacy path (<code>path</code> doesn't start with <code>$</code>):<br>
+     *           Returns the length of the new array after appending <code>values</code> to the array
+     *           at <code>path</code>. If multiple paths are matched, returns the last updated array.
+     *           If the JSON value at <code>path</code> is not a array or if <code>path</code> doesn't
+     *           exist, an error is raised. If <code>key</code> doesn't exist, an error is raised.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", "$", "{\"a\": 1, \"b\": [\"one\", \"two\"]}").get();
+     * var res = Json.arrappend(client, gs("doc"), gs("$.b"), new GlideString[] {gs("\"three\"")}).get();
+     * assert Arrays.equals((Object[]) res, new int[] {3}); // New length of the array after appending
+     * res = Json.arrappend(client, gs("doc"), gs(".b"), new GlideString[] {gs("\"four\"")}).get();
+     * assert res.equals(4); // New length of the array after appending
+     * }</pre>
+     */
+    public static CompletableFuture<Object> arrappend(
+            @NonNull BaseClient client,
+            @NonNull GlideString key,
+            @NonNull GlideString path,
+            @NonNull GlideString[] values) {
+        return executeCommand(
+                client, new ArgsBuilder().add(gs(JSON_ARRAPPEND)).add(key).add(path).add(values).toArray());
     }
 
     /**
@@ -874,6 +958,194 @@ public class Json {
     }
 
     /**
+     * Deletes the JSON document stored at <code>key</code>.
+     *
+     * @param client The Valkey GLIDE client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @return The number of elements deleted. 0 if the key does not exist.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", "{\"a\": 1, \"nested\": {\"a\": 2, \"b\": 3}");
+     * Long result = Json.del(client, "doc").get();
+     * assertEquals(result, 1L);
+     * }</pre>
+     */
+    public static CompletableFuture<Long> del(@NonNull BaseClient client, @NonNull String key) {
+        return executeCommand(client, new String[] {JSON_DEL, key});
+    }
+
+    /**
+     * Deletes the JSON document stored at <code>key</code>.
+     *
+     * @param client The Valkey GLIDE client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @return The number of elements deleted. 0 if the key does not exist.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", "{\"a\": 1, \"nested\": {\"a\": 2, \"b\": 3}");
+     * Long result = Json.del(client, gs("doc")).get();
+     * assertEquals(result, 1L);
+     * }</pre>
+     */
+    public static CompletableFuture<Long> del(@NonNull BaseClient client, @NonNull GlideString key) {
+        return executeCommand(client, new GlideString[] {gs(JSON_DEL), key});
+    }
+
+    /**
+     * Deletes the JSON value at the specified <code>path</code> within the JSON document stored at <code>key</code>.
+     *
+     * @param client The Valkey GLIDE client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @param path Represents the path within the JSON document where the value will be deleted.
+     * @return The number of elements deleted. 0 if the key does not exist, or if the JSON path is invalid or does not exist.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", "{\"a\": 1, \"nested\": {\"a\": 2, \"b\": 3}");
+     * Long result = Json.del(client, "doc", "$..a").get();
+     * assertEquals(result, 2L);
+     * }</pre>
+     */
+    public static CompletableFuture<Long> del(
+            @NonNull BaseClient client, @NonNull String key, @NonNull String path) {
+        return executeCommand(client, new String[] {JSON_DEL, key, path});
+    }
+
+    /**
+     * Deletes the JSON value at the specified <code>path</code> within the JSON document stored at <code>key</code>.
+     *
+     * @param client The Valkey GLIDE client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @param path Represents the path within the JSON document where the value will be deleted.
+     * @return The number of elements deleted. 0 if the key does not exist, or if the JSON path is invalid or does not exist.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", "{\"a\": 1, \"nested\": {\"a\": 2, \"b\": 3}");
+     * Long result = Json.del(client, gs("doc"), gs("$..a")).get();
+     * assertEquals(result, 2L);
+     * }</pre>
+     */
+    public static CompletableFuture<Long> del(
+            @NonNull BaseClient client, @NonNull GlideString key, @NonNull GlideString path) {
+        return executeCommand(client, new GlideString[] {gs(JSON_DEL), key, path});
+    }
+
+    /**
+     * Deletes the JSON document stored at <code>key</code>.
+     *
+     * @param client The Valkey GLIDE client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @return The number of elements deleted. 0 if the key does not exist.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", "{\"a\": 1, \"nested\": {\"a\": 2, \"b\": 3}");
+     * Long result = Json.forget(client, "doc").get();
+     * assertEquals(result, 1L);
+     * }</pre>
+     */
+    public static CompletableFuture<Long> forget(@NonNull BaseClient client, @NonNull String key) {
+        return executeCommand(client, new String[] {JSON_FORGET, key});
+    }
+
+    /**
+     * Deletes the JSON document stored at <code>key</code>.
+     *
+     * @param client The Valkey GLIDE client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @return The number of elements deleted. 0 if the key does not exist.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", "{\"a\": 1, \"nested\": {\"a\": 2, \"b\": 3}");
+     * Long result = Json.forget(client, gs("doc")).get();
+     * assertEquals(result, 1L);
+     * }</pre>
+     */
+    public static CompletableFuture<Long> forget(
+            @NonNull BaseClient client, @NonNull GlideString key) {
+        return executeCommand(client, new GlideString[] {gs(JSON_FORGET), key});
+    }
+
+    /**
+     * Deletes the JSON value at the specified <code>path</code> within the JSON document stored at <code>key</code>.
+     *
+     * @param client The Valkey GLIDE client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @param path Represents the path within the JSON document where the value will be deleted.
+     * @return The number of elements deleted. 0 if the key does not exist, or if the JSON path is invalid or does not exist.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", "{\"a\": 1, \"nested\": {\"a\": 2, \"b\": 3}");
+     * Long result = Json.forget(client, "doc", "$..a").get();
+     * assertEquals(result, 2L);
+     * }</pre>
+     */
+    public static CompletableFuture<Long> forget(
+            @NonNull BaseClient client, @NonNull String key, @NonNull String path) {
+        return executeCommand(client, new String[] {JSON_FORGET, key, path});
+    }
+
+    /**
+     * Deletes the JSON value at the specified <code>path</code> within the JSON document stored at <code>key</code>.
+     *
+     * @param client The Valkey GLIDE client to execute the command.
+     * @param key The <code>key</code> of the JSON document.
+     * @param path Represents the path within the JSON document where the value will be deleted.
+     * @return The number of elements deleted. 0 if the key does not exist, or if the JSON path is invalid or does not exist.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", "{\"a\": 1, \"nested\": {\"a\": 2, \"b\": 3}");
+     * Long result = Json.forget(client, gs("doc"), gs("$..a")).get();
+     * assertEquals(result, 2L);
+     * }</pre>
+     */
+    public static CompletableFuture<Long> forget(
+            @NonNull BaseClient client, @NonNull GlideString key, @NonNull GlideString path) {
+        return executeCommand(client, new GlideString[] {gs(JSON_FORGET), key, path});
+    }
+
+    /**
+     * Toggles a Boolean value stored at the root within the JSON document stored at <code>key</code>.
+     *
+     * @param client The client to execute the command.
+     * @param key The key of the JSON document.
+     * @return Returns the toggled boolean value at the root of the document, or <code>null</code> for
+     *     JSON values matching the root that are not boolean. If <code>key</code> doesn't exist,
+     *     returns <code>null</code>.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", true).get();
+     * var res = Json.toggle(client, "doc").get();
+     * assert res.equals(false);
+     * res = Json.toggle(client, "doc").get();
+     * assert res.equals(true);
+     * }</pre>
+     */
+    public static CompletableFuture<Boolean> toggle(@NonNull BaseClient client, @NonNull String key) {
+        return executeCommand(client, new String[] {JSON_TOGGLE, key});
+    }
+
+    /**
+     * Toggles a Boolean value stored at the root within the JSON document stored at <code>key</code>.
+     *
+     * @param client The client to execute the command.
+     * @param key The key of the JSON document.
+     * @return Returns the toggled boolean value at the root of the document, or <code>null</code> for
+     *     JSON values matching the root that are not boolean. If <code>key</code> doesn't exist,
+     *     returns <code>null</code>.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", ".", true).get();
+     * var res = Json.toggle(client, gs("doc")).get();
+     * assert res.equals(false);
+     * res = Json.toggle(client, gs("doc")).get();
+     * assert res.equals(true);
+     * }</pre>
+     */
+    public static CompletableFuture<Boolean> toggle(
+            @NonNull BaseClient client, @NonNull GlideString key) {
+        return executeCommand(client, new ArgsBuilder().add(gs(JSON_TOGGLE)).add(key).toArray());
+    }
+  
+    /**
      * Multiplies the JSON value(s) at the specified <code>path</code> by <code>number</code> within the JSON document stored at <code>key</code>.
      *
      * @param client The client to execute the command.
@@ -953,6 +1225,73 @@ public class Json {
             client,
             new GlideString[] {gs(JSON_NUMMULTBY), key, path, gs(Long.toString(number))}
         );
+    }
+  
+    /**
+     * Toggles a Boolean value stored at the specified <code>path</code> within the JSON document
+     * stored at <code>key</code>.
+     * 
+     * @param client The client to execute the command.
+     * @param key The key of the JSON document.
+     * @param path The path within the JSON document.
+     * @return
+     *     <ul>
+     *       <li>For JSONPath (<code>path</code> starts with <code>$</code>):<br>
+     *           Returns a <code>Boolean[]</code> with the toggled boolean value for every possible
+     *           path, or <code>null</code> for JSON values matching the path that are not boolean.
+     *       <li>For legacy path (<code>path</code> doesn't start with <code>$</code>):<br>
+     *           Returns the value of the toggled boolean in <code>path</code>. If <code>path</code>
+     *           doesn't exist or the value at <code>path</code> isn't a boolean, an error is raised.
+     *     </ul>
+     *     If <code>key</code> doesn't exist, returns <code>null</code>.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", "$", "{\"bool\": true, \"nested\": {\"bool\": false, \"nested\": {\"bool\": 10}}}").get();
+     * var res = Json.toggle(client, "doc", "$..bool").get();
+     * assert Arrays.equals((Boolean[]) res, new Boolean[] {false, true, null});
+     * res = Json.toggle(client, "doc", "bool").get();
+     * assert res.equals(true);
+     * var getResult = Json.get(client, "doc", "$").get();
+     * assert getResult.equals("{\"bool\": true, \"nested\": {\"bool\": true, \"nested\": {\"bool\": 10}}}");
+     * }</pre>
+     */
+    public static CompletableFuture<Object> toggle(
+            @NonNull BaseClient client, @NonNull String key, @NonNull String path) {
+        return executeCommand(client, new String[] {JSON_TOGGLE, key, path});
+    }
+
+    /**
+     * Toggles a Boolean value stored at the specified <code>path</code> within the JSON document
+     * stored at <code>key</code>.
+     *
+     * @param client The client to execute the command.
+     * @param key The key of the JSON document.
+     * @param path The path within the JSON document.
+     * @return
+     *     <ul>
+     *       <li>For JSONPath (<code>path</code> starts with <code>$</code>):<br>
+     *           Returns a <code>Boolean[]</code> with the toggled boolean value for every possible
+     *           path, or <code>null</code> for JSON values matching the path that are not boolean.
+     *       <li>For legacy path (<code>path</code> doesn't start with <code>$</code>):<br>
+     *           Returns the value of the toggled boolean in <code>path</code>. If <code>path</code>
+     *           doesn't exist or the value at <code>path</code> isn't a boolean, an error is raised.
+     *     </ul>
+     *     If <code>key</code> doesn't exist, returns <code>null</code>.
+     * @example
+     *     <pre>{@code
+     * Json.set(client, "doc", "$", "{\"bool\": true, \"nested\": {\"bool\": false, \"nested\": {\"bool\": 10}}}").get();
+     * var res = Json.toggle(client, gs("doc"), gs("$..bool")).get();
+     * assert Arrays.equals((Boolean[]) res, new Boolean[] {false, true, null});
+     * res = Json.toggle(client, gs("doc"), gs("bool")).get();
+     * assert res.equals(true);
+     * var getResult = Json.get(client, "doc", "$").get();
+     * assert getResult.equals("{\"bool\": true, \"nested\": {\"bool\": true, \"nested\": {\"bool\": 10}}}");
+     * }</pre>
+     */
+    public static CompletableFuture<Object> toggle(
+            @NonNull BaseClient client, @NonNull GlideString key, @NonNull GlideString path) {
+        return executeCommand(
+                client, new ArgsBuilder().add(gs(JSON_TOGGLE)).add(key).add(path).toArray());
     }
 
     /**