+++ python/python/glide/__init__.py
@@ -49,6 +49,11 @@ from glide.async_commands.server_modules.ft_options.ft_create_options import (
     VectorFieldAttributesHnsw,
     VectorType,
 )
+from glide.async_commands.server_modules.ft_options.ft_search_options import (
+    FtSeachOptions,
+    FtSearchLimit,
+    ReturnField,
+)
 from glide.async_commands.sorted_set import (
     AggregationType,
     GeoSearchByBox,
@@ -99,6 +104,7 @@ from glide.config import (
 from glide.constants import (
     OK,
     TOK,
+    FtInfoResponse,
     TClusterResponse,
     TEncodable,
     TFunctionListResponse,
@@ -165,6 +171,7 @@ __all__ = [
     "TResult",
     "TXInfoStreamFullResponse",
     "TXInfoStreamResponse",
+    "FtInfoResponse",
     # Commands
     "BitEncoding",
     "BitFieldGet",
@@ -265,4 +272,7 @@ __all__ = [
     "VectorFieldAttributesFlat",
     "VectorFieldAttributesHnsw",
     "VectorType",
+    "FtSearchLimit",
+    "ReturnField",
+    "FtSeachOptions",
 ]